Refactor pipeline package structure in parser IT
